* .automerge: New.

git-svn-id: file:///srv/svn/kernel1/branches/1.5@1625 1e37f2df-8004-0410-ba96-adcf012cb36e

+++ src/memorymanager/memorymanager.hh
@@ -6,7 +6,7 @@
  */
 namespace MemoryManager
 {
-  extern int allocatedMemory; ///< Allocated total
+  extern size_t allocatedMemory; ///< Allocated total
   /* later
      void * alloc(int sz);
      void  free(void * );
